+++ src/main/java/org/openlmis/auth/service/ExpirationTokenNotifier.java
@@ -26,7 +26,6 @@ import org.openlmis.auth.i18n.ExposedMessageSource;
 import org.openlmis.auth.repository.ExpirationTokenRepository;
 import org.openlmis.auth.service.notification.NotificationService;
 import org.openlmis.auth.service.referencedata.UserReferenceDataService;
-import org.openlmis.util.NotificationRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,8 +33,6 @@ import org.springframework.context.i18n.LocaleContextHolder;
 
 public abstract class ExpirationTokenNotifier<T extends ExpirationToken> {
   public static final long TOKEN_VALIDITY_HOURS = 12;
-
-  static final String MAIL_ADDRESS = System.getenv("MAIL_ADDRESS");
 
   private final Logger logger = LoggerFactory.getLogger(getClass());
 
@@ -67,8 +64,8 @@ public abstract class ExpirationTokenNotifier<T extends ExpirationToken> {
     return repository.save(creator.apply(user));
   }
 
-  void sendEmail(User user, String email, ExpirationToken token,
-      String subjectKey, String bodyKey, String bodyUrl) {
+  void sendEmail(User user, ExpirationToken token, String subjectKey, String bodyKey,
+      String bodyUrl) {
     UserMainDetailsDto referenceDataUser = userReferenceDataService.findOne(user.getId());
 
     String[] bodyMsgArgs = {
@@ -79,11 +76,11 @@ public abstract class ExpirationTokenNotifier<T extends ExpirationToken> {
     String[] subjectMsgArgs = {};
 
     try {
-      notificationService.send(new NotificationRequest(
-          MAIL_ADDRESS,
-          email,
+      notificationService.notify(
+          user,
           messageSource.getMessage(subjectKey, subjectMsgArgs, LocaleContextHolder.getLocale()),
-          messageSource.getMessage(bodyKey, bodyMsgArgs, LocaleContextHolder.getLocale())));
+          messageSource.getMessage(bodyKey, bodyMsgArgs, LocaleContextHolder.getLocale())
+      );
     } catch (Exception exp) {
       logger.warn("Can't send request to the notification service", exp);
       throw new ServerException(exp, ERROR_SEND_NOTIFICATION_FAILURE);